// backend/src/user/dto/update-user.dto.ts
export class UpdateUserDto {
    profilePhoto?: string;
    firstName?: string;
    lastName?: string;
    dob?: Date;
    occupation?: string;
    gender?: string;
    email?: string;
    phoneNumber?: string;
    fax?: string;
    linkedInUrl?: string;
    address?: string;
    city?: string;
    state?: string;
    country?: string;
    zipCode?: string;
    schools?: string[];
  }